+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - `[expect]` Have `Inverse` exportable ([#15714](https://github.com/jestjs/jest/pull/15714))
 - `[expect]` feat: support `async functions` in `toBe` ([#15704](https://github.com/jestjs/jest/pull/15704))
+- `[jest-snapshot]` Handle line endings in snapshots ([#15708](https://github.com/jestjs/jest/pull/15708))
 
 ## 30.0.3